+++ src/questions/questions.service.ts
@@ -1,9 +1,9 @@
 import { PrismaService } from "src/database/prisma.service";
 import { Questions } from "./interface/questions.interface";
-import { Injectable, ParseIntPipe } from "@nestjs/common";
+import { HttpStatus, Injectable, ParseIntPipe } from "@nestjs/common";
 import { CreateQuestionsDto } from "./dto/create-questions.dto";
 import { UpdateQuestionsDto } from "./dto/update-questions.dto";
-
+import { HttpException } from "@nestjs/common";
 
 @Injectable()
 export class QuestionsServices {
@@ -28,17 +28,30 @@ export class QuestionsServices {
   }
 
   
-  async updateQuestions(id:number, updateQuestionsDto: UpdateQuestionsDto): Promise<Questions> {
+  async updateQuestions(id:number, updateQuestionsDto: UpdateQuestionsDto, user_id_request: number): Promise<Questions> {
+    if (id !== user_id_request){
+      throw new HttpException(
+        'Você não tem acesso, pois essa pergunta nao é sua.',
+        HttpStatus.UNAUTHORIZED,
+
+      )}
     return await this.prisma.questions.update({
       where: { id }, 
       data: updateQuestionsDto, 
     });
   }
 
-  async deleteQuestions(id: number): Promise<Questions> {
-    return await this.prisma.questions.delete({
+  async deleteQuestions(id: number, user_id_request: number): Promise<{message:string}> {
+    if (id !== user_id_request){
+      throw new HttpException(
+        'Você não tem acesso, pois essa pergunta nao é sua.',
+        HttpStatus.UNAUTHORIZED,
+
+      )}
+    await this.prisma.questions.delete({
       where: { id },
     });
+    return {message: "Pergunta deletada!"}
   }
